+++ inc/ti/fn/fnisbool.h
@@ -16,3 +16,9 @@ static int do__f_is_bool(ti_query_t * query, cleri_node_t * nd, ex_t * e)
 
     return e->nr;
 }
+
+static int do__f_isbool(ti_query_t * query, cleri_node_t * nd, ex_t * e)
+{
+    log_debug("function `isbool` is deprecated, use `is_bool` instead");
+    return do__f_is_bool(query, nd, e);
+}