+++ inc/ti/fn/fnislist.h
@@ -16,3 +16,9 @@ static int do__f_is_list(ti_query_t * query, cleri_node_t * nd, ex_t * e)
 
     return e->nr;
 }
+
+static int do__f_islist(ti_query_t * query, cleri_node_t * nd, ex_t * e)
+{
+    log_debug("function `islist` is deprecated, use `is_list` instead");
+    return do__f_is_list(query, nd, e);
+}