+++ inc/ti/fn/fnisnil.h
@@ -16,3 +16,9 @@ static int do__f_is_nil(ti_query_t * query, cleri_node_t * nd, ex_t * e)
 
     return e->nr;
 }
+
+static int do__f_isnil(ti_query_t * query, cleri_node_t * nd, ex_t * e)
+{
+    log_debug("function `isnil` is deprecated, use `is_nil` instead");
+    return do__f_is_nil(query, nd, e);
+}